refactor(naming):  AddComment->CommentInput&CommentArray -> CommentBox&CommentGrid -> CommentColumn으로 이름 수정#189
